+++ samples/ruby/lambdas.rb
@@ -1,0 +1,16 @@
+
+# No Space
+l = ->name { puts "Hi #{name}!" }
+
+# 1 Arg
+l = -> one { ... }
+
+# 2 Args
+l = -> one, two { ... }
+
+# 3 Args
+l = -> one, two, three { ... }
+
+# 4 Args
+l = -> one, two, three, four { ... }
+