+++ samples/rust/method.rs
@@ -5,7 +5,7 @@ struct Circle {
 }
 
 impl Circle {
-    fn area(&self) -> f64 {
+    // fn area(&self) -> f64 {
         std::f64::consts::PI * (self.radius * self.radius)
     }
 }