+++ samples/rust/strings.rs
@@ -1,0 +1,3 @@
+"Hello!"
+
+let chara = ':'